move rag.py to shared utilities
